Remove keymap parsing code from jquery.js
