+++ src/store/modules/keymap.js
@@ -7,6 +7,7 @@ import colorways from '@/components/colorways';
 import defaults from './config';
 import axios from 'axios';
 import { backend_skeletons_url } from './constants';
+import { parseKeycode } from './parse.js';
 
 const state = {
   keymap: [[]], // array of arrays
@@ -184,6 +185,33 @@ const actions = {
       .catch(err => {
         console.warn('unable to get keymap template. error:', err);
       });
+  },
+  //Function that takes in a keymap loops over it and fills populates the keymap variable
+  load_converted_keymap({ commit }, converted_keymap) {
+    const store = this;
+    //Loop over each layer from the keymap
+    console.log('converted_keymap', converted_keymap);
+    const acc = converted_keymap.reduce(
+      (acc, layerData, _layer) => {
+        //Add layer object for every layer that exists
+        commit('initLayer', { layer: _layer });
+        //Loop over each keycode in the layer
+        acc.layers.push(
+          layerData.map(keycode => {
+            return parseKeycode(store, keycode, acc.stats);
+          })
+        );
+        acc.stats.layers += 1;
+        return acc;
+      },
+      {
+        stats: { count: 0, any: 0, layers: 0 },
+        layers: []
+      }
+    );
+    commit('setLayers', acc.layers);
+    console.log('stat', acc.stats);
+    return acc.stats;
   }
 };
 const mutations = {
